requirements.txt
transformers
datasets
accelerate
peft
gradio
pandas
tokenizers
openmind
swanlab
loguru
bitsandbytes
python-dotenv
openai
